Normalize non-canonical casts deprecated on PHP 8.5

+++ admin/modules/tools/spamlog.php
@@ -235,7 +235,7 @@ if(!$mybb->get_input('action'))
 		{
 			if(isset($data['confidence']))
 			{
-				$confidence = (double)$data['confidence'].'%';
+				$confidence = (float)$data['confidence'].'%';
 			}
 		}
